Add files via upload
In summary, while the model performs well in predicting customers who did not purchase, it struggles to accurately predict customers who did purchase. This could indicate a class imbalance or other issues in the dataset, and further analysis and model refinement may be necessary to improve predictions for the positive class.